CameraPrefab, Researchs, Utiliy and TerrainToPlane
Commiting the new Camera Prefab, Researchs folder with testing code, new
Utility scripts with the TerrainToPlane Script:

Creates an Plane with format of the terrain. Part of the Fog of War to
D20 Prototype. Still, not the best solution, since there is a limit of
size and vertices. Need to be altered to handle those cases (creating
more planes, etc)